splitting files from Skeleton
