+++ src/Cooked/MockChain/UtxoState.hs
@@ -6,6 +6,7 @@ module Cooked.MockChain.UtxoState
     UtxoPayloadSet (..),
     UtxoPayload (..),
     holdsInState,
+    SkelContext (..),
   )
 where
 
@@ -64,3 +65,10 @@ instance Monoid UtxoPayloadSet where
 -- | Computes the total value in a set
 utxoPayloadSetTotal :: UtxoPayloadSet -> Api.Value
 utxoPayloadSetTotal = mconcat . fmap utxoPayloadValue . utxoPayloadSet
+
+-- | The missing information on a 'TxSkel' that can only be resolved by querying
+-- the state of the blockchain.
+data SkelContext = SkelContext
+  { skelContextTxOuts :: Map Api.TxOutRef Api.TxOut,
+    skelContextTxSkelOutDatums :: Map Api.DatumHash TxSkelOutDatum
+  }
